# swagger.json
{"openapi":"3.0.0","info":{"title":"API Books","version":"1.0.0","description":"This is a simple CRUD API","license":{"name":"MIT","url":"http://mit.com"},"contact":{"name":"Francesc Pallares","url":"https://github.com/fpallaresa","email":"francescpallares@email.com"}},"server":[{"url":"http://localhost:3000"}],"paths":{"/author":{"get":{"summary":"Lists all authors","tags":["Author"],"parameters":[{"in":"query","name":"page","schema":{"type":"integer"},"description":"Page number"},{"in":"query","name":"limit","schema":{"type":"integer"},"description":"Number of items per page"}],"responses":{"200":{"description":"The list of authors","content":{"application/json":{"schema":{"type":"object","properties":{"pagination":{"type":"object","$ref":"#/components/schemas/Pagination","properties":{"totalItems":{"type":"integer"},"totalPages":{"type":"integer"},"currentPage":{"type":"integer"}}},"data":{"type":"array","items":{"$ref":"#/components/schemas/Author"}}}}}}}}},"post":{"summary":"Create a new author","tags":["Author"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Author"}}}},"responses":{"201":{"description":"The author was created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Author"}}}}}}},"/author/name/{name}":{"get":{"summary":"Get an author by name","tags":["Author"],"parameters":[{"in":"path","name":"name","schema":{"type":"string"},"required":true,"description":"The author's name"}],"responses":{"200":{"description":"The author info","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Author"}}}}},"404":{"description":"The author was not found"}}}},"/author/{id}":{"put":{"summary":"Update an author","tags":["Author"],"parameters":[{"in":"path","name":"id","schema":{"type":"string"},"required":true,"description":"The author ID"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Author"}}}},"responses":{"200":{"description":"The author was updated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Author"}}}},"404":{"description":"The author was not found"}}},"delete":{"summary":"Delete an author","tags":["Author"],"parameters":[{"in":"path","name":"id","schema":{"type":"string"},"required":true,"description":"The author ID"}],"responses":{"200":{"description":"The author was deleted successfully"},"404":{"description":"The author was not found"}}},"get":{"summary":"Get an author by ID","tags":["Author"],"parameters":[{"in":"path","name":"id","schema":{"type":"string"},"required":true,"description":"The author ID"},{"in":"query","name":"includeBooks","schema":{"type":"boolean"},"description":"Whether to include the author's books"}],"responses":{"200":{"description":"The author info","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Author"}}}},"404":{"description":"The author was not found"}}}},"/author/image-upload":{"post":{"summary":"Upload a profile image for an author","tags":["Author"],"consumes":["multipart/form-data"],"parameters":[{"in":"formData","name":"file","type":"file","description":"The file to upload."},{"in":"formData","name":"authorId","type":"string","description":"The id of the author"}],"responses":{"200":{"description":"The image was uploaded successfully"},"404":{"description":"The author was not found"}}}},"/author/login":{"post":{"summary":"Iniciar sesión como autor","tags":["Author"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","description":"Credenciales de inicio de sesión del autor","required":true,"schema":{"type":"object","properties":{"email":{"type":"string","format":"email","description":"Email del autor"},"password":{"type":"string","format":"password","description":"Contraseña del autor"}}}}],"responses":{"200":{"description":"Autor autenticado exitosamente","schema":{"type":"object","properties":{"token":{"type":"string","description":"Token de acceso JWT generado para el autor autenticado"}}}},"400":{"description":"Campos de email y/o contraseña faltantes o incorrectos"},"401":{"description":"Email y/o contraseña incorrectos"}}}},"/book":{"get":{"summary":"Lists all the books","tags":["Book"],"parameters":[{"in":"query","name":"page","schema":{"type":"integer","minimum":1},"description":"Page number"},{"in":"query","name":"limit","schema":{"type":"integer","minimum":1},"description":"Number of items per page"}],"responses":{"200":{"description":"The list of books","schema":{"type":"object","properties":{"totalItems":{"type":"integer","description":"Total number of books"},"totalPages":{"type":"integer","description":"Total number of pages"},"currentPage":{"type":"integer","description":"Current page"},"data":{"type":"array","items":{"$ref":"#/definitions/Book"}}}},"content":{"application/json":{"schema":{"type":"object","properties":{"pagination":{"type":"object","properties":{"totalItems":{"type":"integer"},"totalPages":{"type":"integer"},"currentPage":{"type":"integer"}}},"data":{"type":"array","items":{"$ref":"#/components/schemas/Book"}}}}}}},"400":{"description":"Invalid page or limit parameters"}}},"post":{"summary":"Create a new book","tags":["Book"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/definitions/Book"}}}},"responses":{"201":{"description":"The book was created successfully","schema":{"$ref":"#/definitions/Book"}},"500":{"description":"Internal server error"}}}},"/book/title/{title}":{"get":{"summary":"Get a book by title","tags":["Book"],"parameters":[{"in":"path","name":"title","schema":{"type":"string"},"required":true,"description":"The title of the book"}],"responses":{"200":{"description":"The book info","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/definitions/Book"}}}}},"404":{"description":"The book with the specified title was not found"}}}},"/book/{id}":{"put":{"summary":"Update a book","tags":["Book"],"parameters":[{"in":"path","name":"id","schema":{"type":"string"},"required":true,"description":"The ID of the book"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/definitions/Book"}}}},"responses":{"200":{"description":"The book was updated successfully","content":{"application/json":{"schema":{"$ref":"#/definitions/Book"}}}},"404":{"description":"The book with the specified ID was not found"}}},"get":{"summary":"Get a book by ID","tags":["Book"],"parameters":[{"in":"path","name":"id","schema":{"type":"string"},"required":true,"description":"The ID of the book"}],"responses":{"200":{"description":"The book information","content":{"application/json":{"schema":{"$ref":"#/definitions/Book"}}}},"404":{"description":"The book with the specified ID was not found"}}}},"/file-upload":{"post":{"summary":"Upload a file","tags":["File Upload"],"consumes":["multipart/form-data"],"parameters":[{"in":"formData","name":"file","type":"file","description":"The file to upload."}],"responses":{"200":{"description":"The file was uploaded successfully"}}}}},"components":{"schemas":{"Author":{"type":"object","required":["email","password","name"],"properties":{"email":{"type":"string","description":"Email address of the author (sample author@example.com)"},"password":{"type":"string","description":"Password of the author (at least 8 characters)"},"name":{"type":"string","description":"Name of the author (sample John Doe)"},"country":{"type":"string","description":"Country of the author (sample SPAIN)","enum":["SPAIN","ITALY","USA","GERMANY","JAPAN","ENGLAND","COLOMBIA","RUSSIA","UNITED STATES","ARGENTINA"]},"profileImage":{"type":"string","description":"Profile image URL of the author (sample http://example.com/image.jpg)"},"books":{"type":"array","items":{"$ref":"#/components/schemas/Book"},"description":"List of books written by the author"}}},"Book":{"type":"object","required":["title"],"properties":{"title":{"type":"string","description":"Title of the book (sample My Book Title)"},"author":{"type":"string","description":"ID of the author (sample 60d0fe4f5311236168a109ca)"},"pages":{"type":"number","description":"Number of pages in the book (sample 300)"},"publisher":{"type":"object","properties":{"name":{"type":"string","description":"Name of the publisher (sample Penguin Books)"},"country":{"type":"string","description":"Country of the publisher (sample USA)","enum":["SPAIN","ITALY","USA","GERMANY","JAPAN"]}}}}},"Error":{"type":"object","properties":{"error":{"type":"string","description":"The error message"}}},"Pagination":{"type":"object","properties":{"totalItems":{"type":"number","description":"Total items on this endpoint"},"totalPages":{"type":"number","description":"Number of pages for this pagination"},"currentPage":{"type":"number","description":"Current page showed"}}}}},"tags":[{"name":"Author","description":"API endpoints for authors"},{"name":"Book","description":"API endpoints for books"},{"name":"File Upload","description":"API endpoints for file upload"}]}